enable fat lto on release and optimizations on dev
fat lto significantly improves seek performance, but
it compiles so slowly that is annoying to use for
debugging and testing.
the dev build on the other hand compiles fast, but
runs very slowly and more optimizations don't slow
down the building process significantly, but greatly
improve runtime performance.

i might need to adjust the dev profile, once i merge
mpris and dbus support, as that may be noticeably
slower for build times with optimizations.